# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/wiki/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wiki"
                ],
                "summary": "Создание элемента энциклопедии",
                "responses": {
                    "200": {
                        "description": "Ok\" //хз как тут сделать",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseOk"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/wiki/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wiki"
                ],
                "summary": "Получение элементов энциклопедии",
                "parameters": [
                    {
                        "description": "Параметры выборки",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.listEncyclopediaItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseWithDetails-handler_listEncyclopediaItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/wiki/{EncyclopediaItemId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wiki"
                ],
                "summary": "Получение элемента энциклопедии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор элемента энциклопедии",
                        "name": "EncyclopediaItemId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseWithDetails-handler_getEncyclopediaItemResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/shttp.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.getEncyclopediaItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "settlementId": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handler.listEncyclopediaItemElement": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "settlement_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handler.listEncyclopediaItemRequest": {
            "type": "object",
            "properties": {
                "settlement_id": {
                    "type": "string"
                }
            }
        },
...
        "shttp.ResponseError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {}
            }
        },
        "shttp.ResponseOk": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {}
            }
        },
...
        "shttp.ResponseWithDetails-handler_getEncyclopediaItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/handler.getEncyclopediaItemResponse"
                }
            }
        },
...
        "shttp.ResponseWithDetails-handler_listEncyclopediaItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.listEncyclopediaItemElement"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ExternalId": {
            "type": "apiKey",
            "name": "ExternalId",
            "in": "header"
        }
    }
}